refactor: Auth 모듈 분리 (#23) (KAN-76)
