+++ src/main/java/org/ioteatime/meonghanyangserver/auth/controller/AuthController.java
@@ -1,10 +1,11 @@
-package org.ioteatime.meonghanyangserver.user.controller;
+package org.ioteatime.meonghanyangserver.auth.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.ioteatime.meonghanyangserver.auth.dto.request.SendEmailRequest;
+import org.ioteatime.meonghanyangserver.auth.service.AuthService;
 import org.ioteatime.meonghanyangserver.common.api.Api;
 import org.ioteatime.meonghanyangserver.user.dto.UserDto;
-import org.ioteatime.meonghanyangserver.user.service.JoinService;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,12 +14,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/open-api/auth")
-public class JoinController implements JoinApi {
-    private final JoinService joinService;
+public class AuthController implements AuthApi {
+    private final AuthService authService;
 
     @PostMapping("/sign-up")
     public Api<Object> registerUser(@Valid @RequestBody UserDto userDto) {
-        joinService.joinProcess(userDto);
+        authService.joinProcess(userDto);
         return Api.CREATE();
+    }
+
+    @PostMapping("/email-verification")
+    public Api<?> verifyEmail(@Valid @RequestBody SendEmailRequest sendEmailReq) {
+        authService.send(sendEmailReq.email());
+        return Api.OK();
     }
 }
