+++ src/main/java/org/ioteatime/meonghanyangserver/auth/dto/request/SendEmailRequest.java
@@ -1,10 +1,10 @@
-package org.ioteatime.meonghanyangserver.mail.dto.request;
+package org.ioteatime.meonghanyangserver.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
 
-public record SendRequest(
+public record SendEmailRequest(
         @Valid @Email @NotNull @Schema(description = "이메일", example = "example@gmail.com")
                 String email) {}
